feat: EduHarness 可审计 AI 助教（Agent 引擎 × 教育闭环）
- 会话/作答/事件全量落库，SSE+WebSocket 断线续传，可回放可恢复
- 教师/学生多角色：自研 HMAC Token 鉴权、归属隔离、工具审批桥接
- 可申诉评阅：作答沉淀证据链，掌握度纯函数幂等重算，改判自动联动
- 艾宾浩斯 1/2/4/7/15/30 复习调度 + LeetCode 训练营与 Python 入门双课程
- 五层端到端评测与发布门禁；40 项分层用例 + 引擎跨平台 CI

+++ minicode/cost_tracker.py
@@ -183,10 +183,10 @@ def calculate_cost(
     """
     pricing = MODEL_PRICING.get(model, MODEL_PRICING["default"])
     return (
-        (input_tokens / _DECIMAL_1M) * pricing["input"]
-        + (output_tokens / _DECIMAL_1M) * pricing["output"]
-        + (cache_read_tokens / _DECIMAL_1M) * pricing["cache_read"]
-        + (cache_creation_tokens / _DECIMAL_1M) * pricing["cache_write"]
+        float(input_tokens / _DECIMAL_1M) * pricing["input"]
+        + float(output_tokens / _DECIMAL_1M) * pricing["output"]
+        + float(cache_read_tokens / _DECIMAL_1M) * pricing["cache_read"]
+        + float(cache_creation_tokens / _DECIMAL_1M) * pricing["cache_write"]
     )
